for when only the view= is defined in the metrics
